Add option to limit osc max stiffness force/torque
Adds two extra member variables to OSCImpendanceController for
configuring max translational-force/rotational-torque generated
from the impedance stiffness. These parameters are used to clip
the L2 norm of the stiffness force/torques in the franka control
loop. Also augment the protobuff message used for this controller
so that these values can be configured from the python side.